+++ src/components/eye/deployChildren/basicDeploy/trashDeploy.vue
@@ -133,28 +133,35 @@
                     this.choseInfoId.push(id)
                 }
                 if (this.choseInfoId.length > 0) {
-                    api.dustbin.deleteDustbin(this.choseInfoId).then(res => {
-                        console.log(res, '删除成功')
-                        for (let i = 0; i < this.choseInfoId.length; i++) {
-                            this.trashList = this.trashList.filter((item, index) => {
-                                if (item.dustbinBean.id === this.choseInfoId[i]){
-                                    this.trashList[index].checked = false
-                                    this.trashList[index].status = false
-                                }
-                                return item.status === true
-                            })
-                        }
-                        this.$message.success('删除成功')
-                        this.choseInfoId = []
-                    }).catch(err => {
-                        console.log(err)
-                        this.$message.error('删除失败，请稍后重试')
-                        this.choseInfoId = []
+                    this.$confirm('此操作将永久删除该数据, 是否继续?', '提示', {
+                        confirmButtonText: '确定',
+                        cancelButtonText: '取消',
+                        type: 'warning'
+                    }).then(() => {
+                        api.dustbin.deleteDustbin(this.choseInfoId).then(res => {
+                            console.log(res, '删除成功')
+                            for (let i = 0; i < this.choseInfoId.length; i++) {
+                                this.trashList = this.trashList.filter((item, index) => {
+                                    if (item.dustbinBean.id === this.choseInfoId[i]){
+                                        this.trashList[index].checked = false
+                                        this.trashList[index].status = false
+                                    }
+                                    return item.status === true
+                                })
+                            }
+                            this.$message.success('删除成功')
+                            this.choseInfoId = []
+                        }).catch(err => {
+                            console.log(err)
+                            this.$message.error('删除失败，请稍后重试')
+                            this.choseInfoId = []
+                        })
+                    }).catch(() => {
+                        this.$message.info('取消删除')
                     })
                 }else {
-                    this.$message.error('请选择要删除的选项')
+                    this.$message.error('请选择要删除的垃圾桶信息')
                 }
-
             },
             toggleList (type) {
                 if (type === 'list') {
@@ -237,6 +244,9 @@
                     this.$message.success('修改成功')
                     this.choseInfoId = []
                     this.getAllTrash()
+                }).catch(err => {
+                    console.log(err, '修改失败')
+                    this.$message.error('修改失败，请稍后重试')
                 })
             },
             addNewTrash (info) {
@@ -255,7 +265,11 @@
                 console.log(trashObj, 'this is trashObj')
                 api.dustbin.createDustbin(JSON.stringify(trashObj)).then(res => {
                     console.log('增加成功')
+                    this.$message.success('创建成功')
                     this.getAllTrash()
+                }).catch(err => {
+                    console.log(err,'创建失败')
+                    this.$message.error('创建失败，请稍后重试')
                 })
             },
             fixedInfo (id) {
